requirements.txt
apiai==1.2.3
certifi==2018.8.13
chardet==3.0.4
DateTime==4.2
idna==2.7
numpy==1.15.1
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
telegram==0.0.1
urllib3==1.23
zope.interface==4.5.0
